Recoge los valores de la api y los almacena en Realm aunque con algunos valores null

+++ app/src/main/java/com/example/dndcharacterapp/models/character/SkillProficiency.kt
@@ -2,5 +2,5 @@ package com.example.dndcharacterapp.models.character
 
 data class SkillProficiency(
     val name: String? = null,
-    val bonus: Int? = null
+    val bonus: Double? = null
 )
